Revert "Merge #4359"
This reverts commit 6afe56b925e015a15c014d3d2df3e8c4df6575b2, reversing
changes made to 6756c472339ad8a56239e86b5257e39917f584c8.
