+++ node/src/components/consensus/cl_context.rs
@@ -13,7 +13,6 @@ use crate::{
 };
 
 #[derive(DataSize)]
-#[cfg_attr(test, derive(Clone))]
 pub struct Keypair {
     secret_key: Arc<SecretKey>,
     public_key: PublicKey,
